move enum into same file
